Rename sequence inspector to sequenceExpression
