+++ src/components/inspectors/sequenceExpression.js
